Fix conversion warnings with explicit static_cast<>s

+++ src/types/vector.cpp
@@ -7,9 +7,9 @@
 using namespace std;
 
 
-Type* Vector::insert(int index, Type* object) {
+Type* Vector::insert(long int index, Type* object) {
     if (index < 0) { index = (internal_object.size()+index); }
-    if ((index < 0) or (index >= (int)internal_object.size() and internal_object.size() != 0)) {
+    if ((index < 0) or (index >= static_cast<long int>(internal_object.size()) and internal_object.size() != 0)) {
         throw new OutOfRangeException("vector index out of range");
     }
     vector<Type*>::iterator it = (internal_object.begin()+index);
@@ -22,14 +22,14 @@ Type* Vector::push(Type* object) {
     return object;
 }
 
-Type* Vector::pop(int index) {
+Type* Vector::pop(long int index) {
     // FIXME: allow popping from arbitrary indexes
     Type* ptr = internal_object.back();
     internal_object.pop_back();
     return ptr;
 }
 
-Type* Vector::at(int index) {
+Type* Vector::at(long int index) {
     if (index < 0) { index = (internal_object.size()+index); }
     if ((index < 0) or (index >= (int)internal_object.size())) {
         throw new OutOfRangeException("vector index out of range");
